chores: correct error + multi threads
